requirements.txt
argon2-cffi==20.1.0
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.5
dj-database-url==0.5.0
Django==3.2.5
django-ckeditor==6.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.5.1
et-xmlfile==1.1.0
gunicorn==20.1.0
numpy==1.21.0
openpyxl==3.0.7
pandas==1.2.5
Pillow==8.3.1
psycopg2==2.9.1
pycparser==2.20
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.2.0
xlrd==2.0.1
